+++ lib/openapi_spec.dart
@@ -2,3 +2,5 @@
 library openapi_spec;
 
 export 'src/open_api/index.dart';
+export 'src/generators/index.dart'
+    show SchemaGeneratorOptions, ClientGeneratorOptions, ServerGeneratorOptions;