+++ lib/src/generators/index.dart
@@ -9,7 +9,11 @@ import 'package:openapi_spec/src/open_api/index.dart';
 import 'package:recase/recase.dart';
 
 part 'base.dart';
+part 'keywords.dart';
 part 'client.dart';
+part 'client_options.dart';
 part 'schema.dart';
+part 'schema_options.dart';
 part 'server.dart';
+part 'server_options.dart';
 part 'validation.dart';