+++ lib/src/generators/server_options.dart
@@ -1,0 +1,13 @@
+part of openapi_generators;
+
+// ==========================================
+// CLASS: ServerGeneratorOptions
+// ==========================================
+
+class ServerGeneratorOptions extends BaseGeneratorOptions {
+  const ServerGeneratorOptions({
+    super.enabled = false,
+    super.replaceOutput,
+    super.includeVersion,
+  });
+}